Move EULA prompt

+++ MCBEgetZIP.sh
@@ -25,6 +25,17 @@ installed_ver=$(ls ~/bedrock-server-[0-9]*\.[0-9]*\.[0-9]*\.[0-9]*\.zip 2> /dev/
 
 if ! echo "$installed_ver" | grep -q "$current_ver"; then
 # There might be more than one ZIP in ~
+	echo Enter Y if you agree to the Minecraft End User License Agreement and Privacy Policy
+	echo Minecraft End User License Agreement: https://minecraft.net/terms
+	# Does prompting the EULA seem so official that it violates the EULA?
+	echo Privacy Policy: https://go.microsoft.com/fwlink/?LinkId=521839
+	read -r input
+	input=$(echo "$input" | tr '[:upper:]' '[:lower:]')
+	if [ "$input" != y ]; then
+		>&2 echo "$input != y"
+		exit 3
+	fi
+
 	trap 'sudo rm -f ~/"$current_ver"' ERR
 	wget --prefer-family=IPv4 "$url" -O ~/"$current_ver"
 	trap - ERR
